Settings: Change navbar general behaviour (2/2)
This commit changes that navbar is present on all devices even
with hardware keys. On hardware key devices the navbar height is
set to 0 as default now. Several reasons for that.

- no SystemUI restart when de/activating (last one we had :p)
- our navbar is very resource friendly so even no problem for the
  low end devices
- we can now use some of the navbar stuff you will see in next commits
  navbar magic :p

This commit also respects the different handling off menu's. Thats
why a part of the code (hasNavigationBar()) will be still left because
it is needed in main for

core/java/android/view/ViewConfiguration.java
core/java/android/view/IWindowManager.aidl
core/java/android/view/WindowManagerPolicy.java

this files which are controling that.

Settings commit:
- additional cleanup of no needed imports

cheers

Change-Id: Ida80c6ecaeaf9f20fff80a58f2b62c30d2ba54d6

Conflicts:
	src/com/android/settings/slim/LockscreenWidgets.java
	src/com/android/settings/slim/NotificationDrawerSettings.java
	src/com/android/settings/slim/quicksettings/QuickSettingsTilesStyle.java

Conflicts:

	src/com/android/settings/cyanogenmod/NavbarSettings.java