Settings: dslv do not allow to add recent, home, back main action twice
to have it twice causes problems on navbar pie and wherever in future
home, back, recents is used when the system disables this buttons and
more then one per feature exists. Due that a user do not set eg
2 home buttons on navbar let us block it and give the user a hint
that this is not allowed. All other actions can be added twice

Change-Id: I729d166139403e65b948f45acf36a47b7de7eb4c